#vfb38 Split into smaller components
